% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_method_for_ratio.R
\name{DeltaMethodForRatio}
\alias{DeltaMethodForRatio}
\title{The Delta Method for Ratio}
\description{
The Delta Method for Ratio

The Delta Method for Ratio
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeltaMethodForRatio-new}{\code{DeltaMethodForRatio$new()}}
\item \href{#method-DeltaMethodForRatio-get_expected_value}{\code{DeltaMethodForRatio$get_expected_value()}}
\item \href{#method-DeltaMethodForRatio-get_variance}{\code{DeltaMethodForRatio$get_variance()}}
\item \href{#method-DeltaMethodForRatio-get_squared_standard_error}{\code{DeltaMethodForRatio$get_squared_standard_error()}}
\item \href{#method-DeltaMethodForRatio-get_standard_error}{\code{DeltaMethodForRatio$get_standard_error()}}
\item \href{#method-DeltaMethodForRatio-get_confidence_interval}{\code{DeltaMethodForRatio$get_confidence_interval()}}
\item \href{#method-DeltaMethodForRatio-get_info}{\code{DeltaMethodForRatio$get_info()}}
\item \href{#method-DeltaMethodForRatio-clone}{\code{DeltaMethodForRatio$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeltaMethodForRatio-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeltaMethodForRatio-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeltaMethodForRatio$new(numerator, denominator, bias_correction = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeltaMethodForRatio-get_expected_value"></a>}}
\if{latex}{\out{\hypertarget{method-DeltaMethodForRatio-get_expected_value}{}}}
\subsection{Method \code{get_expected_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeltaMethodForRatio$get_expected_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeltaMethodForRatio-get_variance"></a>}}
\if{latex}{\out{\hypertarget{method-DeltaMethodForRatio-get_variance}{}}}
\subsection{Method \code{get_variance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeltaMethodForRatio$get_variance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeltaMethodForRatio-get_squared_standard_error"></a>}}
\if{latex}{\out{\hypertarget{method-DeltaMethodForRatio-get_squared_standard_error}{}}}
\subsection{Method \code{get_squared_standard_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeltaMethodForRatio$get_squared_standard_error()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeltaMethodForRatio-get_standard_error"></a>}}
\if{latex}{\out{\hypertarget{method-DeltaMethodForRatio-get_standard_error}{}}}
\subsection{Method \code{get_standard_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeltaMethodForRatio$get_standard_error()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeltaMethodForRatio-get_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-DeltaMethodForRatio-get_confidence_interval}{}}}
\subsection{Method \code{get_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeltaMethodForRatio$get_confidence_interval(
  alternative = "two.sided",
  conf_level = 0.95
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeltaMethodForRatio-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-DeltaMethodForRatio-get_info}{}}}
\subsection{Method \code{get_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeltaMethodForRatio$get_info(alternative = "two.sided", conf_level = 0.95)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeltaMethodForRatio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeltaMethodForRatio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeltaMethodForRatio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
