% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltatest.R
\name{deltatest}
\alias{deltatest}
\title{Z-Test Using the Delta Method}
\usage{
deltatest(
  data,
  formula,
  by,
  group_names = "auto",
  type = c("difference", "relative_change"),
  bias_correction = TRUE,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  na.rm = FALSE,
  quiet = FALSE
)
}
\description{
Z-Test Using the Delta Method
}
