% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_data.R
\name{generate_dummy_data}
\alias{generate_dummy_data}
\title{Generate Dummy Data}
\usage{
generate_dummy_data(
  n_user,
  model = c("Bernoulli", "normal"),
  sigma = 0,
  xi = 0,
  random_unit = c("user", "session", "page"),
  treatment_ratio = 0.5
)
}
\description{
Generate Dummy Data
}
